decorator for posting json request to vault #issue:2 ,user update token generates and save new token to user instance
