+++ token_auth/json_status.py
@@ -1,0 +1,8 @@
+from __future__ import unicode_literals
+from django.utils.translation import ugettext_lazy as _
+
+STATUS_OK = _('OK')
+STATUS_FAIL = _('Failed')
+STATUS_USER_DOES_NOT_EXIST = _('No such user')
+STATUS_INVALID_TOKEN = _('Bad authentication')
+
